resolve comments and fully remove instabug
