+++ android/app/build.gradle
@@ -147,7 +147,6 @@ android {
 }
 
 dependencies {
-    compile project(':instabug-reactnative')
     compile project(':react-native-device-info')
     compile (project(':react-native-camera')) {
         exclude group: "com.google.android.gms"
