+++ android/app/src/main/java/lab/childmindinstitute/data/MainApplication.java
@@ -3,7 +3,6 @@ package lab.childmindinstitute.data;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
-import com.instabug.reactlibrary.RNInstabugReactnativePackage;
 import com.learnium.RNDeviceInfo.RNDeviceInfo;
 import org.reactnative.camera.RNCameraPackage;
 import com.sensors.RNSensorsPackage;
@@ -37,25 +36,19 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-            		new RNInstabugReactnativePackage.Builder("de12937dca290605e0a66f106b5921bf",MainApplication.this)
-							.setInvocationEvent("screenshot")
-							.setPrimaryColor("#1D82DC")
-							.setFloatingEdge("left")
-							.setFloatingButtonOffsetFromTop(250)
-							.build(),
-            new RNDeviceInfo(),
-            new RNCameraPackage(),
-            new RNSensorsPackage(),
-            new ImagePickerPackage(),
-            new ReactNativePushNotificationPackage(),
-            new SvgPackage(),
-            new AudioPackage(),
-            new RNSoundPackage(),
-            new OrientationPackage(),
-            new OGWavePackage(),
-            new ReactNativeAudioPackage(),
-            new VectorIconsPackage(),
-            new RNFetchBlobPackage()
+          new RNDeviceInfo(),
+          new RNCameraPackage(),
+          new RNSensorsPackage(),
+          new ImagePickerPackage(),
+          new ReactNativePushNotificationPackage(),
+          new SvgPackage(),
+          new AudioPackage(),
+          new RNSoundPackage(),
+          new OrientationPackage(),
+          new OGWavePackage(),
+          new ReactNativeAudioPackage(),
+          new VectorIconsPackage(),
+          new RNFetchBlobPackage()
       );
     }
