+++ app/themes/activityTheme.js
@@ -1,0 +1,68 @@
+import {StyleSheet} from 'react-native';
+
+const styles=StyleSheet.create({
+  body: {
+    flex: 1,
+  },
+  camera: {
+    width: '100%',
+    height: 360,
+    position: 'relative',
+  },
+  footer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    padding: 15,
+  },
+  footerText: {
+    fontSize: 20,
+    fontWeight: '300',
+  }, buttonText: {
+    fontSize: 24,
+    fontWeight: '300'
+  },
+  takeButton: {
+    borderRadius: 12,
+    width: '100%',
+    height: 360,
+    alignContent: 'center',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderWidth: 4,
+    borderColor: '#d10000',
+    backgroundColor: '#ffdddd'
+  },
+  videoConfirmed: {
+    borderRadius: 12,
+    width: '100%',
+    height: 360,
+    alignContent: 'center',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderWidth: 4,
+    borderColor: '#00a30a',
+    backgroundColor: '#99ff9f'
+  },
+  chooseButton: {
+    borderRadius: 12,
+    width: '100%',
+    height: 116,
+    alignContent: 'center',
+    justifyContent: 'center',
+    alignItems: 'center',
+    borderWidth: 4,
+    borderColor: '#000000',
+    backgroundColor: '#dbdbdb',
+    marginTop: 8,
+  },
+  redIcon: {
+    color: '#d10000',
+    fontSize: 60,
+  },
+  greenIcon: {
+    color: '#00a30a',
+    fontSize: 60,
+  }
+});
+
+export default styles
